// src/lib/api/model/inboxMessagesPageDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lean Commerce API
 * The AI Sales Agent CRM API Documentation
 * OpenAPI spec version: 1.0
 */
import type { InboxMessageDto } from "./inboxMessageDto";

export interface InboxMessagesPageDto {
  data: InboxMessageDto[];
  hasMore: boolean;
  /** @nullable */
  nextCursor: string | null;
}